requirements.txt
# Core web framework - NO SCIPY DEPENDENCIES!
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation  
pydantic==2.5.0
pydantic-settings==2.1.0

# Database (dual SQLite/PostgreSQL support)
sqlalchemy==2.0.23
greenlet==3.2.4
# SQLite async support
aiosqlite==0.19.0  
# PostgreSQL support
asyncpg==0.29.0

# HTTP and LLM
httpx==0.25.2
requests==2.31.0
openai==1.3.8
tenacity==8.2.3

# Utilities
click==8.1.7
rich==13.7.0
python-dotenv==1.0.0
markdown==3.4.4

# Testing (basic only)
pytest==7.4.3
pytest-asyncio==0.21.1

# Analysis and visualization packages (for compare.sh)
numpy==2.1.1
matplotlib==3.9.2
pandas==2.2.2
scikit-learn==1.5.2
seaborn==0.13.2
plotly==5.22.0

# NOTE: scipy, alembic, quart, redis removed to avoid build issues
# Install manually if needed: pip install scipy redis alembic quartmatplotlib>=3.7.0
numpy>=1.24.0
seaborn>=0.12.0
pillow>=9.0.0
